backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0

# RAG dependencies
sentence-transformers>=2.2.0
pinecone>=3.0.0
groq>=0.4.0

# RAGAS Evaluation
ragas>=0.2.0
datasets>=2.14.0
